+++ app/Http/Controllers/BackEnd/CustomerReviewController.php
@@ -47,6 +47,11 @@ class CustomerReviewController extends Controller
         $review->name = $request->fullname;
         $review->content = $request->reviewContent;
         if($request->file('image') != null){
+            $checkFile = self::CheckFileUpload($request->file('image')->getClientOriginalName());
+            if(!$checkFile){
+                $request->session()->flash('fail','Image format is invalid (jpg,jpeg,png,gif)!');
+                return redirect()->back();
+            }
             $image = $request->file('image')->store('public/reviews');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
@@ -87,6 +92,11 @@ class CustomerReviewController extends Controller
         $review->name = $request->fullname;
         $review->content = $request->reviewContent;
         if($request->file('image') != null){
+            $checkFile = self::CheckFileUpload($request->file('image')->getClientOriginalName());
+            if(!$checkFile){
+                $request->session()->flash('fail','Image format is invalid (jpg,jpeg,png,gif)!');
+                return redirect()->back();
+            }
             if($review->image != 'avatar.png'){
                 //Xoa anh cu~
                 File::delete('storage/reviews/'.$review->image);
@@ -126,5 +136,13 @@ class CustomerReviewController extends Controller
             $request->session()->flash('fail','Delete unsuccessfully!');
         }
         return redirect()->back();
+    }
+    public static function CheckFileUpload($filename){
+        $arrFilename = explode('.',$filename);
+        $format = $arrFilename[count($arrFilename)-1];
+        if ($format == 'png' || $format == 'jpg' ||$format == 'jpeg' ||$format == 'gif' ){
+            return true;
+        }
+        return false;
     }
 }
