+++ app/Http/Controllers/BackEnd/GiftCardController.php
@@ -30,6 +30,11 @@ class GiftCardController extends Controller
         $giftCard = new GiftCard();
         $giftCard->title = $request->title;
         if($request->file('image') != null){
+            $checkFile = self::CheckFileUpload($request->file('image')->getClientOriginalName());
+            if(!$checkFile){
+                $request->session()->flash('fail','Image format is invalid (jpg,jpeg,png,gif)!');
+                return redirect()->back();
+            }
             $image = $request->file('image')->store('public/gift-card');
             $arr_filename = explode("/",$image);
             $filename = end($arr_filename);
@@ -56,6 +61,11 @@ class GiftCardController extends Controller
         $giftCard = GiftCard::findOrFail($id);
         $giftCard->title = $request->title;
         if($request->file('image') != null){
+            $checkFile = self::CheckFileUpload($request->file('image')->getClientOriginalName());
+            if(!$checkFile){
+                $request->session()->flash('fail','Image format is invalid (jpg,jpeg,png,gif)!');
+                return redirect()->back();
+            }
             if($giftCard->image != 'default.png'){
                 //Xoa anh cu~
                 File::delete('storage/gift-card/'.$giftCard->image);
@@ -88,5 +98,13 @@ class GiftCardController extends Controller
             $request->session()->flash('fail','Delete unsuccessfully!');
         }
         return redirect()->back();
+    }
+    public static function CheckFileUpload($filename){
+        $arrFilename = explode('.',$filename);
+        $format = $arrFilename[count($arrFilename)-1];
+        if ($format == 'png' || $format == 'jpg' ||$format == 'jpeg' ||$format == 'gif' ){
+            return true;
+        }
+        return false;
     }
 }
